fix(frontend): リノートの判定が甘いのを修正
